pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jay <joaomarcelo1015ja@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.5"
djangorestframework = "^3.15.2"
pillow = "^11.1.0"
pytube = "^15.0.0"
pytubefix = "^8.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
